# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nuke Vehicle Data API
  description: |
    The Nuke API provides programmatic access to vehicle data management, observations,
    and real-time event notifications via webhooks.

    ## Authentication

    All API requests require authentication using either:

    - **API Key**: Pass via `X-API-Key` header
    - **Bearer Token**: Supabase JWT via `Authorization: Bearer <token>` header

    Get your API key at [https://nuke.com/settings/api-keys](https://nuke.com/settings/api-keys)

    ## Rate Limiting

    API requests are rate limited. When you exceed your limit, you'll receive a 429 response.
    Check the `X-RateLimit-Remaining` header to monitor your usage.

    ## Webhooks

    Subscribe to real-time events by creating webhook endpoints. Events are delivered with
    HMAC-SHA256 signatures for verification.
  version: 1.0.0
  contact:
    name: Nuke Support
    url: https://nuke.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://qkgaybvrernstplzjaam.supabase.co/functions/v1
    description: Production API
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /webhooks-manage/{id}:
    get:
      tags: [Webhooks]
      summary: Retrieve webhook endpoint
      operationId: getWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook endpoint with recent deliveries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/WebhookEndpoint'
                      - type: object
                        properties:
                          recent_deliveries:
                            type: array
                            items:
                              $ref: '#/components/schemas/WebhookDelivery'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Webhooks]
      summary: Update webhook endpoint
      operationId: updateWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdateInput'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/WebhookEndpoint'
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Webhooks]
      summary: Delete webhook endpoint
      operationId: deleteWebhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /webhooks-manage/{id}/rotate-secret:
    post:
      tags: [Webhooks]
      summary: Rotate webhook secret
      description: Generate a new signing secret. The old secret will be invalidated immediately.
      operationId: rotateWebhookSecret
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Secret rotated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/WebhookEndpoint'
                      - type: object
                        properties:
                          secret:
                            type: string
                  message:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key from https://nuke.com/settings/api-keys
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase session JWT
  schemas:
...
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        description:
          type: string
          nullable: true
        events:
          type: array
          items:
            type: string
            enum:
              - '*'
              - vehicle.created
              - vehicle.updated
              - vehicle.deleted
              - observation.created
              - document.uploaded
              - import.completed
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        stats:
          type: object
          properties:
            total_deliveries:
              type: integer
            failed_deliveries:
              type: integer
    WebhookDelivery:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_type:
          type: string
        event_id:
          type: string
        status:
          type: string
          enum: [pending, success, failed, retrying]
        attempts:
          type: integer
        response_status:
          type: integer
          nullable: true
        response_time_ms:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        delivered_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
...
    WebhookUpdateInput:
      type: object
      properties:
        url:
          type: string
          format: uri
        description:
          type: string
        events:
          type: array
          items:
            type: string
        is_active:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Vehicle not found
...